+++ src/Grphp/Client.php
@@ -129,6 +129,7 @@ class Client
             $i->setMethod($method);
             $i->setMetadata($metadata);
             $i->setOptions($options);
+            $i->setStub($this->client);
             return $i->call(function() use (&$i, &$interceptors, &$method, &$request, &$metadata, &$options, &$callback) {
                 $metadata = $i->getMetadata();
                 $request = $i->getRequest();